+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Remove upper version bound on Django, to allow upgrading to Django 4.1.1 bugfix release.
+* Handle oauthlib errors on create token requests
 
 ## [2.1.0] 2022-06-19